+++ src/js/lib/card-nicknames/vow.ts
@@ -1,0 +1,123 @@
+// Crimson Vow (VOW)
+export default [
+  {
+    // https://scryfall.com/card/vow/3/cathar-squad-glow-up?utm_source=api
+    setCode: "vow",
+    collectorNumber: "3",
+    realName: ["Arm the Cathars"],
+    nickname: ["Cathar Squad Glow Up"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/20/candlestick-breaker?utm_source=api
+    setCode: "vow",
+    collectorNumber: "20",
+    realName: ["Hopeful Initiate"],
+    nickname: ["Candlestick Breaker"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/32/resistance-squad?utm_source=api
+    setCode: "vow",
+    collectorNumber: "32",
+    realName: ["Resistance Squad"],
+    nickname: ["Frontline Team"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/44/voice-of-the-blessed?utm_source=api
+    setCode: "vow",
+    collectorNumber: "44",
+    realName: ["Voice of the Blessed"],
+    nickname: ["Bedsheet Elemental"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/62/gutter-skulker-gutter-shortcut?utm_source=api
+    setCode: "vow",
+    collectorNumber: "62",
+    realName: ["Gutter Skulker", "Gutter Shortcut"],
+    nickname: ["Reachy Thingie", "Grabby Stuff"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/65/magnifying-glass-enthusiast-enthusiasts-magnifying-glass?utm_source=api
+    setCode: "vow",
+    collectorNumber: "65",
+    realName: ["Jacob Hauken, Inspector", "Hauken's Insight"],
+    nickname: ["Magnifying Glass Enthusiast", "Enthusiast's Magnifying Glass"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/110/edgars-awakening?utm_source=api
+    setCode: "vow",
+    collectorNumber: "110",
+    realName: ["Edgar's Awakening"],
+    nickname: ["Edgar's Good Soup"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/133/undead-butler?utm_source=api
+    setCode: "vow",
+    collectorNumber: "133",
+    realName: ["Undead Butler"],
+    nickname: ["Undead Stewart"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/148/master-of-riposte?utm_source=api
+    setCode: "vow",
+    collectorNumber: "148",
+    realName: ["Cemetery Gatekeeper"],
+    nickname: ["Master of Riposte"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/162/spikey-friend-spikey-pugstomper?utm_source=api
+    setCode: "vow",
+    collectorNumber: "162",
+    realName: ["Ill-Tempered Loner", "Howlpack Avenger"],
+    nickname: ["Spikey Friend", "Spikey Pugstomper"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/169/phantom-elemental?utm_source=api
+    setCode: "vow",
+    collectorNumber: "169",
+    realName: ["Magma Pummeler"],
+    nickname: ["Phantom Elemental"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/172/sanguine-bridesmaids?utm_source=api
+    setCode: "vow",
+    collectorNumber: "172",
+    realName: ["Olivia's Attendants"],
+    nickname: ["Sanguine Bridesmaids"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/202/nick-cages-final-form?utm_source=api
+    setCode: "vow",
+    collectorNumber: "202",
+    realName: ["Hiveheart Shaman"],
+    nickname: ["Nick Cage's Final Form"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/247/sigardian-paladin?utm_source=api
+    setCode: "vow",
+    collectorNumber: "247",
+    realName: ["Sigardian Paladin"],
+    nickname: ["Trainer Trainee"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/vow/255/house-beneviento?utm_source=api
+    setCode: "vow",
+    collectorNumber: "255",
+    realName: ["Dollhouse of Horrors"],
+    nickname: ["House Beneviento"],
+    source: "Scryfall Preview Name",
+  },
+];